massive switch to using 1-base column indexing
Consume, return, and utilize 1-base indexing for columns in a Data table, instead of VERY old (and outdated) precedent of 0-base indexing. This has been on a todo list for very long time, and WILL BREAK legacy code. But dealing with column indexes of 0 is awful: is it true, is it not true but defined?, is it just empty? Enabling warnings makes it worse. Fortunately, switching is fairly painless. We simply include a null value as the 0th column.
Also, simplify and reduce code for add_xxxx_metadata methods while I'm at it.